13_Rag/requirement.txt
youtube-transcript-api
langchain-community
langchain-openai
faiss-cpu
tiktoken
python-dotenv